requirements.txt
django==2.2.6
djangorestframework
django-filter
pandas
pillow
requests
